Simplify variable names
